pyproject.toml
[project]
name = "evo-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "evolutionapi>=0.0.9",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
